v6_platform/fastapi_backend/requirements.txt
fastapi
uvicorn
requests
pydantic
google-search-results
google-generativeai
python-dotenv
